Factor out `image`
